Checking demo build into source control for use with Github Pages
Formerly the demo build directory was `demo/` and excluded from source
control.

However it needs to be checked in to enable use with GitHub Pages, and
it needs to be named `docs/` since that's the only folder naming
convention that Pages supports.

This build needs to be regenerated after each `[CHANGE]` commit by
re-running `yarn run build`